requirements.txt
# Core ML/DL dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0

# Face analysis (InsightFace)
insightface>=0.7.3
onnxruntime>=1.14.0

# Web app interface
streamlit>=1.30.0

# Jupyter support (for the .ipynb files)
jupyterlab>=4.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Optional visualization and utils
scikit-learn>=1.3.0
pandas>=2.0.0

# Miscellaneous
protobuf<5.0.0
